featured,search,nearby,saved farms api

+++ database/migrations/2024_10_29_113212_create_farms_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users');
             $table->string('name');
             $table->string('location');
-            $table->integer('lat');
-            $table->integer('lng');
+            $table->float('lat');
+            $table->float('lng');
             $table->string('phone');
             $table->string('email');
             $table->string('website');
